creted update function

+++ src/main/kotlin/com/vannak/tech/api_project/repository/RoleRepository.kt
@@ -4,6 +4,6 @@ import com.vannak.tech.api_project.domain.model.Role
 import com.vannak.tech.api_project.domain.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RoleRepository: JpaRepository<Role,Int> {
+interface RoleRepository: JpaRepository<Role,Long> {
 
 }
